Add definition of method koma_can_go of class Gameboard
Add definition of method koma_can_go of class Gameboard
But we meet a problem that we can't take protect member can_go and
direct of subclass Jump and Rush.
We need to add a vitrul method into class Koma and implement into Jump
and Rush to get infomation we need.