Add a method if_koma_can_move_in into gameboard.
Add a method if_koma_can_move_in into gameboard.
Which is used to testing if the Koma weather if it can move in
(board_x,board_y) on board.
This method can use to simply method koma_can_put.
And use in testing weather if koma must to promote or not after move.