Edit protected method koma_go in Gameboard
Edit protected method koma_go in Gameboard.
Now we can initial gameboard, move koma, and drop koma.
After this, we need to check whether player is win the game or not.
To do this, we need to check whether Ou is been checkmate( by check
whether is check around the Ou or not) or not.
So todo list is check->checkmate->is_win